Commit organization work, folders
