+++ client/src/components/Your Outfit/AddOutfitButton.jsx
@@ -1,0 +1,39 @@
+import React, { useState, useEffect } from 'react';
+import axios from 'axios';
+import { makeStyles } from '@material-ui/core/styles';
+import Card from '@material-ui/core/Card';
+import CardActions from '@material-ui/core/CardActions'
+import CardContent from '@material-ui/core/CardContent';
+import Button from '@material-ui/core/Button';
+import Typography from '@material-ui/core/Typography';
+
+const useStyles = makeStyles({
+  root: {
+    minWidth: 275,
+  },
+  title: {
+    fontSize: 14,
+  },
+  pos: {
+    marginBottom: 12,
+  }
+});
+
+
+const AddOutfitButton = () => {
+  const classes = useStyles();
+
+  return (
+    <React.Fragment>
+      <Card className={classes.root} variant="outlined">
+        <CardContent>
+          <Typography className={classes.title} gutterBottom>
+            I am an Outfit Button! I will have a button attached!
+          </Typography>
+        </CardContent>
+      </Card>
+    </React.Fragment>
+  );
+}
+
+export default AddOutfitButton;
