--- client/src/components/YourOutfit.jsx
@@ -1,7 +1,0 @@
-import React, { useState } from 'react';
-
-const YourOutfit = () => {
-  return <div>This will be the Outfit component!</div>;
-};
-
-export default YourOutfit;
